benchmarks/longmemeval/requirements.txt
datasets>=2.14.0
litellm>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
